+++ src/framework/core/Shape.js
@@ -1,7 +1,7 @@
 import * as THREE from 'three';
 
 import {loadMaterial, extend} from '../extras/api';
-import {Loop} from '../extensions/Loop';
+import {Loop} from '../utils/Loop';
 import {World} from './World';
 import {CoreObject} from './CoreObject';